requirements.txt
# Core Dependencies
numpy>=1.24.3
pandas>=2.1.4
tqdm>=4.66.1
python-dotenv>=1.0.0

# PDF Processing (Unstructured.io + Tesseract)
unstructured[pdf]>=0.10.0
pytesseract>=0.3.10
Pillow>=10.1.0

# Legacy (keeping for compatibility)
docling>=1.0.0

# PDF Processing Fallback
PyMuPDF>=1.23.8
easyocr>=1.7.0

# Embeddings and Models
sentence-transformers>=2.6.0
accelerate>=0.25.0
# BAAI/bge-m3 specific dependencies
torch>=2.1.0
transformers>=4.45.0

# Direct Transformers (Qwen2.5-7B)
torch>=2.1.0
transformers>=4.45.0
accelerate>=0.25.0

# RAG Framework
llama-index>=0.10.0
langchain>=0.2.0
langchain-community>=0.2.0

# Vector Database
faiss-cpu>=1.7.4

# Language Detection
langdetect>=1.0.9

# Text Processing
nltk>=3.8.1
spacy>=3.7.2

# Development
pytest>=7.4.3
black>=23.11.0
flake8>=6.1.0